pinocchio_system

Module instructions

Modules

advance_nonce_account 🔒

allocate 🔒

allocate_with_seed 🔒

assign 🔒

assign_with_seed 🔒

authorize_nonce_account 🔒

create_account 🔒

create_account_with_seed 🔒

initialize_nonce_account 🔒

transfer 🔒

transfer_with_seed 🔒

update_nonce_account 🔒

withdraw_nonce_account 🔒

Structs

AdvanceNonceAccount

Consumes a stored nonce, replacing it with a successor.

Allocate

Allocate space in a (possibly new) account without funding.

AllocateWithSeed

Allocate space for and assign an account at an address derived from a base public key and a seed.

Assign

Assign account to a program

AssignWithSeed

Assign account to a program based on a seed.

AuthorizeNonceAccount

Change the entity authorized to execute nonce instructions on the account.

CreateAccount

Create a new account.

CreateAccountWithSeed

Create a new account at an address derived from a base pubkey and a seed.

InitializeNonceAccount

Drive state of Uninitialized nonce account to Initialized, setting the nonce value.

Transfer

Transfer lamports.

TransferWithSeed

Transfer lamports from a derived address.

UpdateNonceAccount

One-time idempotent upgrade of legacy nonce versions in order to bump them out of chain blockhash domain.

WithdrawNonceAccount

Withdraw funds from a nonce account.